AI calls work, merge v1 done. need UI improvements, other fixes.
